HW7 Oprava 15.12
